update some config file

+++ WebServer/src/com/pitaya/bookingnow/service/IAccountService.java
@@ -8,6 +8,6 @@ import org.springframework.dao.DataAccessException;
 public interface IAccountService<T> {
 	public boolean addAccount(T entity) throws DataAccessException;   
 	public boolean execute(T entity) throws DataAccessException;   
-	public T getAccount(Integer id) throws DataAccessException;   
+	public T getAccount(Long id) throws DataAccessException;   
 	public List<T> getList() throws DataAccessException;
 }
